Handle multiline URLs, including a test case
Hanging indents indicate a continued line, and the indent get stripped.
This is some sort of 'folded' style I think.

I don't remember seeing this in official 'specs' but I've seen actual http
files doing it. Probably headers can be folded as well but I'll cross
that bridge when I get to it.